backend/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
python-dotenv==1.0.0
requests==2.31.0
gunicorn==20.1.0
urllib3<2.0.0
